Refactor generation logic to a new file to avoid circular dependency between raster.py and eeraster.py
