+++ eeraster.py
@@ -281,5 +281,9 @@ def dem():
       'projects/sat-io/open-datasets/GLO-30').select("b1"))
   return _dem
 
-
+# A collection of column names to functions that load the corresponding 
+# earth engine asset.
+column_name_to_ee_asset_fn = {
+  "DEM": dem,
+}
